src/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
base58==2.1.1
bitarray==2.9.2
black==24.3.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cytoolz==0.12.3
distlib==0.3.8
eip712-structs==1.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
exceptiongroup==1.2.1
filelock==3.13.1
flake8==7.0.0
frozenlist==1.4.1
fsspec==2024.2.0
hexbytes==0.3.1
huggingface-hub==0.21.4
identify==2.5.35
idna==3.6
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.3.0
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
multiaddr==0.0.9
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
netaddr==1.2.1
networkx==3.2.1
nodeenv==1.8.0
numpy==1.26.4
packaging==24.0
parsimonious==0.8.1
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.5.0
pre-commit==3.7.0
protobuf==3.19.5
py-clob-client==0.15.1
py_order_utils==0.3.0
pycodestyle==2.11.1
pycryptodome==3.20.0
pyflakes==3.2.0
pysha3==1.0.2
pytest==8.2.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
rlp==2.0.1
rpds-py==0.18.1
safetensors==0.4.2
scipy==1.12.0
six==1.16.0
sympy==1.12
tokenizers==0.15.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
urllib3==2.2.1
varint==1.0.2
virtualenv==20.25.1
web3==5.31.4
websockets==9.1
yarl==1.9.4
